refactor(mobile): rename importPrivateKey to importSigner
